scraper/pcpartpicker/rag-api/pyproject.toml
[project]
name = "rag-api"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "lancedb>=0.26.1",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
